Full materialize function
