+++ sdk/python/feast/feature_store.py
@@ -11,15 +11,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from datetime import datetime
 from pathlib import Path
-from typing import List, Optional, Union
+from typing import Any, Dict, List, Optional, Tuple, Type, Union
 
 import pandas as pd
+import pyarrow
 
 from feast.entity import Entity
 from feast.feature_view import FeatureView
 from feast.infra.provider import Provider, get_provider
-from feast.offline_store import RetrievalJob, get_offline_store_for_retrieval
+from feast.offline_store import OfflineStore, get_offline_store, RetrievalJob, get_offline_store_for_retrieval
 from feast.registry import Registry
 from feast.repo_config import (
     LocalOnlineStoreConfig,
@@ -27,6 +29,8 @@ from feast.repo_config import (
     RepoConfig,
     load_repo_config,
 )
+from feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
+from feast.types.Value_pb2 import Value as ValueProto
 
 
 class FeatureStore:
@@ -57,6 +61,9 @@ class FeatureStore:
 
     def _get_provider(self) -> Provider:
         return get_provider(self.config)
+
+    def _get_offline_store(self) -> Type[OfflineStore]:
+        return get_offline_store(self.config)
 
     def _get_registry(self) -> Registry:
         return Registry(self.config.metadata_store)
@@ -176,3 +183,151 @@ def _get_requested_feature_views(
         feature_views_list.append(view)
 
     return feature_views_list
+
+    @property
+    def project(self) -> str:
+        return "default"
+
+    def materialize(
+        self, feature_views: List[str], start_date: datetime, end_date: datetime
+    ):
+        full_feature_views = []
+        registry = self._get_registry()
+        for name in feature_views:
+            feature_view = registry.get_feature_view(name, self.project)
+            full_feature_views.append(feature_view)
+            if feature_view.input.table_ref is None:
+                raise NotImplementedError(
+                    f"This function is only implemented for FeatureViews with a table_ref; {feature_view.name} does not have one."
+                )
+
+        # TODO paging large loads
+        for feature_view in full_feature_views:
+            (
+                entity_names,
+                feature_names,
+                event_timestamp_column,
+                created_timestamp_column,
+            ) = _run_reverse_field_mapping(feature_view)
+
+            offline_store = self._get_offline_store()
+            table = offline_store.pull_latest_from_table(
+                feature_view.input.table_ref,
+                entity_names,
+                feature_names,
+                event_timestamp_column,
+                created_timestamp_column,
+                start_date,
+                end_date,
+            )
+
+            if feature_view.input.field_mapping is not None:
+                table = _run_forward_field_mapping(
+                    table, feature_view.input.field_mapping
+                )
+
+            rows_to_write = _convert_arrow_to_proto(table, feature_view)
+
+            provider = self._get_provider()
+            provider.online_write_batch(
+                "default", feature_view, rows_to_write, created_timestamp_column
+            )
+
+
+def _run_reverse_field_mapping(
+    feature_view: FeatureView,
+) -> Tuple[List[str], List[str], str, Optional[str]]:
+    """
+    If a field mapping exists, run it in reverse on the entity names, feature names, event timestamp column, and created timestamp column to get the names of the relevant columns in the BigQuery table.
+
+    Args:
+        feature_view: FeatureView object containing the field mapping as well as the names to reverse-map.
+    Returns:
+        Tuple containing the list of reverse-mapped entity names, reverse-mapped feature names, reverse-mapped event timestamp column, and reverse-mapped created timestamp column that will be passed into the query to the offline store.
+    """
+    # if we have mapped fields, use the original field names in the call to the offline store
+    event_timestamp_column = feature_view.input.event_timestamp_column
+    entity_names = [entity for entity in feature_view.entities]
+    feature_names = [feature.name for feature in feature_view.features]
+    created_timestamp_column = feature_view.input.created_timestamp_column
+    if feature_view.input.field_mapping is not None:
+        reverse_field_mapping = {
+            v: k for k, v in feature_view.input.field_mapping.items()
+        }
+        event_timestamp_column = (
+            reverse_field_mapping[event_timestamp_column]
+            if event_timestamp_column in reverse_field_mapping.keys()
+            else event_timestamp_column
+        )
+        created_timestamp_column = (
+            reverse_field_mapping[created_timestamp_column]
+            if created_timestamp_column is not None
+            and created_timestamp_column in reverse_field_mapping.keys()
+            else created_timestamp_column
+        )
+        entity_names = [
+            reverse_field_mapping[col] if col in reverse_field_mapping.keys() else col
+            for col in entity_names
+        ]
+        feature_names = [
+            reverse_field_mapping[col] if col in reverse_field_mapping.keys() else col
+            for col in feature_names
+        ]
+    return (
+        entity_names,
+        feature_names,
+        event_timestamp_column,
+        created_timestamp_column,
+    )
+
+
+def _run_forward_field_mapping(
+    table: pyarrow.Table, field_mapping: Dict[str, str],
+) -> pyarrow.Table:
+    # run field mapping in the forward direction
+    cols = table.column_names
+    mapped_cols = [
+        field_mapping[col] if col in field_mapping.keys() else col for col in cols
+    ]
+    table = table.rename_columns(mapped_cols)
+    return table
+
+
+def _convert_arrow_to_proto(
+    table: pyarrow.Table, feature_view: FeatureView
+) -> List[Tuple[EntityKeyProto, Dict[str, ValueProto], datetime]]:
+    rows_to_write = []
+    for row in zip(*table.to_pydict().values()):
+        entity_key = EntityKeyProto()
+        for entity_name in feature_view.entities:
+            entity_key.entity_names.append(entity_name)
+            idx = table.column_names.index(entity_name)
+            value = _convert_to_proto(row[idx])
+            entity_key.entity_values.append(value)
+        feature_dict = {}
+        for feature in feature_view.features:
+            idx = table.column_names.index(feature.name)
+            value = _convert_to_proto(row[idx])
+            feature_dict[feature.name] = value
+        event_timestamp_idx = table.column_names.index(
+            feature_view.input.event_timestamp_column
+        )
+        rows_to_write.append((entity_key, feature_dict, row[event_timestamp_idx]))
+    return rows_to_write
+
+
+def _convert_to_proto(value: Any) -> ValueProto:
+    value_proto = ValueProto()
+    if isinstance(value, str):
+        value_proto.string_val = value
+    elif isinstance(value, bool):
+        value_proto.bool_val = value
+    elif isinstance(value, int):
+        value_proto.int32_val = value
+    elif isinstance(value, float):
+        value_proto.double_val = value
+    elif isinstance(value, bytes):
+        value_proto.bytes_val = value
+    else:
+        raise ValueError(f"Cannot convert value {value} of type {type(value)}.")
+    return value_proto
